fix: Restuctuing the th app folder     - Moving the indivual component to the one component folder
