+++ src/app/Components/index.ts
@@ -1,0 +1,3 @@
+export { Product } from "./Product";
+export { FAQ } from "./FAQ";
+export { FeatureCard } from "./FeatureCard";
