--- src/app/Home/Components/index.ts
@@ -1,3 +1,0 @@
-export { Product } from "./Product";
-export { FAQ } from "./FAQ";
-export { FeatureCard } from "./FeatureCard";
